requirements.txt
aiohttp
xmltodict
python-dotenv
pandas
matplotlib
git+https://github.com/RichyA/pyit500.git
discord.py
